# server/documentation/swagger-doc-ugram.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "API for the Ugram project (instagram-like) done as part of the GLO-3112 course at Laval University.",
    "version": "1",
    "title": "Ugram API"
  },
  "servers": [
    {
      "url": "http://ugram-team02.pm9h7ckh7u.us-east-2.elasticbeanstalk.com",
      "description": "API-URL"
    }
  ],
...
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "create a new user",
        "description": "",
        "operationId": "addUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "400": {
            "description": "Missing or bad parameter"
          }
        }
      },
      "get": {
        "tags": [
          "users"
        ],
        "summary": "list all user",
        "description": "Get all user",
        "operationId": "getAllUser",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "perPage",
            "in": "query",
            "description": "Number of result",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserComplete"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "delete logged user",
        "description": "Delete logged user",
        "operationId": "deleteUser",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          }
        }
      },
      "UserComplete": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "pictureUrl": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "registrationDate": {
            "type": "object",
            "properties": {
              "afterNow": {
                "type": "boolean"
              },
              "beforeNow": {
                "type": "boolean"
              },
              "equalNow": {
                "type": "boolean"
              }
            }
          }
        }
      },
...
      "Token": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}